+++ monarch_mini/src/framing.rs
@@ -46,10 +46,10 @@ use crate::connection::MonitorOp;
 use crate::connection::SendPayload;
 use crate::connection::SideChannelAction;
 use crate::connection::SideChannelMessage;
+use crate::heartbeat::BeatKind;
+use crate::heartbeat::ConnectionId;
+use crate::heartbeat::Heartbeat;
 use crate::msg::MsgPart;
-use crate::quic_heartbeat::BeatKind;
-use crate::quic_heartbeat::ConnectionId;
-use crate::quic_heartbeat::Heartbeat;
 use crate::shm::MapperHandle;
 use crate::shm::SHM_THRESHOLD;
 use crate::shm::ShmClient;
