+++ monarch_mini/src/lib.rs
@@ -10,12 +10,14 @@ mod actor;
 mod connection;
 mod ctx;
 mod framing;
+mod heartbeat;
 mod inproc_transport;
 mod matcher;
 mod msg;
+mod net;
+mod net_transport;
 mod poller;
-mod quic_heartbeat;
-mod quic_transport;
+mod quic_net;
 mod shm;
 mod transport;
 mod unix_framing;
